nixpkgs 16.09: *backup: use pubring.kbx
https://www.gnupg.org/faq/whats-new-in-2.1.html

If no pubring.gpg is found, gpg defaults to the new keybox
format and creates a pubring.kbx keybox file. If such a
keybox file already exists, for example due to the use of
gpgsm, it will also be used for OpenPGP keys. However, if
a pubring.gpg is found and no keybox file with OpenPGP keys
exists, the old pubring.gpg will be used. Take care: GnuPG
versions before 2.1 will always use the pubring.gpg file
and not know anything about keys stored in the keybox file.